% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trmm2ncdf.R
\name{dim_ncdf}
\alias{dim_ncdf}
\title{dim_ncdf function}
\usage{
dim_ncdf(dat)
}
\arguments{
\item{dat}{an object inheriting class trmmNCDF as created by trmm2ncdf}
}
\value{
the dimension of the NetCDF file
}
\description{
A function to return the dimension of the NetCDF file saved by trmm2ncdf
}
\seealso{
\link{trmm2ncdf}
}

