% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{get_brick}
\alias{get_brick}
\title{read all layers into a brick}
\usage{
get_brick(L)
}
\arguments{
\item{L}{trmmheader object}
}
\value{
raster brick
}
\description{
get all layers into a brick
}
\details{
reads all layers into a brick, with names.
}
\author{
Barry Rowlingson
}
