% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{get_header}
\alias{get_header}
\title{read trmm header}
\usage{
get_header(f)
}
\arguments{
\item{f}{path to trmm file}
}
\value{
a trmmheader object
}
\description{
read header from trmm file
}
\details{
This functions read a header from a trmm file so you can
then read the data into a raster or raster brick.
}
\author{
Barry Rowlingson
}
