% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readdata.R
\name{get_layer}
\alias{get_layer}
\title{read one layer from a trmm file}
\usage{
get_layer(L, layer)
}
\arguments{
\item{L}{trmmlayer object}

\item{layer}{layer variable name}
}
\value{
raster of the layer
}
\description{
get a layer from a trmm using the header
}
\details{
when you have a trmm header object, use this to get a single
layer into a raster object
}
\author{
Barry Rowlingson
}

