% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getTRMM}
\alias{getTRMM}
\title{getTRMM function}
\usage{
getTRMM(outdir, start = lubridate::ymd("20121231"),
  end = lubridate::ymd("20130101"), by = "days", hours = c("00", "03",
  "06", "09", "12", "15", "18", "21"), product = "3B42RT", version = 7,
  website = "ftp://trmmopen.gsfc.nasa.gov/pub/merged")
}
\arguments{
\item{outdir}{the directory in which to save files}

\item{start}{start date e.g. ymd('20121231')}

\item{end}{end date e.g. ymd('20130101')}

\item{by}{argument passed to seq(start,end,by), the sequence of days to download data from, default is all days between start and end}

\item{hours}{the hours in the day for which to download data, a character string. Default is all hours.}

\item{product}{the product to download, default is '3B42RT', but other options are '3B41RT' and '3B40RT'}

\item{version}{version of the data to download, default is 7, but could be either 5 or 6. Note that version 7 has the greatest temporal coverage at present}

\item{website}{the website from which to do the downloading, default is 'ftp://trmmopen.gsfc.nasa.gov/pub/merged'}
}
\value{
Downloads the files requested into the directory specified.
}
\description{
A function to download 3-hourly TRMM data into a directory of your choice.
}

