% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trmm.R
\docType{package}
\name{trmm-package}
\alias{trmm}
\alias{trmm-package}
\title{trmm}
\format{\preformatted{ logi NA
}}
\usage{
trmm
}
\description{
Read data from NASA TRMM data files into raster structures.
}
\details{
\tabular{ll}{
Package: \tab trmm\cr
Version: \tab 0.1\cr
Date: \tab 2015-06-18\cr
License: \tab MIT \cr
}



section{Dependencies}{
The package \code{trmm} depends upon some other important contributions to CRAN in order to operate; their uses here are indicated:\cr\cr
    sp, raster, stringr, lubridate, ncdf.
}

section{Citation}{
X
}

references{
X
}
}
\author{
Barry Rowlingson, Department of Medicine, Lancaster University,
 Benjamin Taylor, Department of Medicine, Lancaster University
}
\keyword{package}

