% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trmm2ncdf.R
\name{trmm2ncdf}
\alias{trmm2ncdf}
\title{trmm2ncdf function}
\usage{
trmm2ncdf(directory, outfile = tempfile(fileext = ".nc"),
  start = ymd("20060101"), end = ymd("20060102"), by = "days",
  hours = c("00", "03", "06", "09", "12", "15", "18", "21"),
  product = "3B42RT", version = 7, poly = NULL, brick = TRUE,
  layer = NULL, quiet = FALSE)
}
\arguments{
\item{directory}{directory in which the .bin files are stored. Note these must be gunzipped first}

\item{outfile}{name and location of the NetCDF file to save results into}

\item{start}{start date e.g. ymd('20121231')}

\item{end}{end date e.g. ymd('20130101')}

\item{by}{argument passed to seq(start,end,by), the sequence of days to download data from, default is all days between start and end}

\item{hours}{the hours in the day for which to download data, a character string. Default is all hours.}

\item{product}{the product to download, default is '3B42RT', but other options are '3B41RT' and '3B40RT'}

\item{version}{version of the data to download, default is 7, but could be either 5 or 6. Note that version 7 has the greatest temporal coverage at present}

\item{poly}{a polygonal window bounding a spatial subset of the data to extract. By default, all data over the range covered by TRMM will be extracted}

\item{brick}{logical. If set to TRUE (the default) then this extracts all TRMM data at each time point}

\item{layer}{the name of the layer to extract. Options are ''}

\item{quiet}{logical: show progress?}
}
\value{
an object of class trmmNCDF, for which there is a method extract_ncdf for extracting data from the saved NetCDF file. The returned list object contains an element firstslice which is a raster brick of data from the first time point, this can be used to give the extracted data geographical context.
}
\description{
A function to convert TRMM rainfall data binaries into a NetCDF file
}

