% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trmm.R
\docType{package}
\name{trmm-package}
\alias{trmm-package}
\alias{trmm}
\title{trmm}
\format{An object of class \code{logical} of length 1.}
\usage{
trmm
}
\description{
Read data from NASA TRMM data files into raster structures.
}
\details{
\tabular{ll}{
Package: \tab trmm\cr
Version: \tab 0.1\cr
Date: \tab 2015-06-18\cr
License: \tab MIT \cr
}



section{Dependencies}{
The package \code{trmm} depends upon some other important contributions to CRAN in order to operate; their uses here are indicated:\cr\cr
    sp, raster, stringr, lubridate, ncdf4.
}

section{Citation}{
X
}

references{
X
}
}
\author{
Barry Rowlingson, Department of Medicine, Lancaster University,
 Benjamin Taylor, Department of Medicine, Lancaster University
}
\keyword{package}
